Adding CODEOWNERS (#2)
By default GitHub does not allow you to specify reviewers of a PR for a
repository that you are not a collaborator on. To work around this we are adding
a CODEOWNERS to the repository to Marisa and Tristan get requested as reviewers
by default for all PRs in this repo.